Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -11,8 +11,8 @@ int main(){
         scanf("%d %d",&rectangles[i].length,&rectangles[i].breadth);
     }
     for(int i=0;i<n;i++){
-        int l=rectangles[i].length;
-        int b=rectangles[i].breadth;
+        float l=rectangles[i].length;
+        float b=rectangles[i].breadth;
         printf("Rectangle %d: Area = %.2f, Perimeter = %.2f\n",i+1,l*b,2*(l+b));
     }
